Fix tests part one. Page 18 is broken, todo fix.

+++ content/geometry/Point.h
@@ -10,7 +10,7 @@
 #pragma once
 
 
-int sgn(long long x) { return (x>0) - (x<0); } // floats compare with eps
+int sgn(ll x) { return (x>0) - (x<0); } // floats compare with eps
 template<class T>
 struct Point {
 	typedef Point P;
